tests/wpt/tests/css/css-ruby/parsing/ruby-merge-invalid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Ruby Layout: parsing ruby-merge with invalid values</title>
<link rel="help" href="https://drafts.csswg.org/css-ruby-1/#ruby-align-property">
<meta name="assert" content="ruby-merge supports only the grammar 'separate | merge | auto'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_invalid_value("ruby-merge", "none");
test_invalid_value("ruby-merge", "collapse");
test_invalid_value("ruby-merge", "10px");

test_invalid_value("ruby-merge", "merge separate");
test_invalid_value("ruby-merge", "merge auto");
test_invalid_value("ruby-merge", "auto separate");
</script>
</body>
</html>
